refactor: moved files around to better match the file structure convention
