--- machine_learning/src/middleware/mod.rs
@@ -1,5 +1,0 @@
-mod param_manager;
-mod server_metadata;
-
-pub use param_manager::{BackIter, FrontIter, ParamManager};
-pub use server_metadata::ServerParamsMetadata;
