+++ machine_learning/src/param_manager.rs
@@ -1,7 +1,32 @@
 use rayon::prelude::*;
 
-use super::ServerParamsMetadata;
 use crate::{MlErr, Result, optimization::Optimizer};
+
+/// The state necessary to make forward and backward passes through the network.
+pub struct ServerParamsMetadata<'mw> {
+    params: &'mw mut [f32],
+    grad: &'mw mut [f32],
+    acc_grad_buf: &'mw mut [f32],
+}
+
+impl<'mw> ServerParamsMetadata<'mw> {
+    /// Creates a new `ServerParamsMetadata`.
+    ///
+    /// # Arguments
+    /// * `params` - The mutable slice of this server's parameters.
+    /// * `grad` - The server's dedicated gradient slice.
+    /// * `acc_grad_buf` - The server's accumulated gradient buffer.
+    ///
+    /// # Returns
+    /// A new `ServerParamsMetadata` instance.
+    pub fn new(params: &'mw mut [f32], grad: &'mw mut [f32], acc_grad_buf: &'mw mut [f32]) -> Self {
+        Self {
+            params,
+            grad,
+            acc_grad_buf,
+        }
+    }
+}
 
 /// The manager of parameters, this middleware's module manages the model's parameter retrieval from the
 /// servers and selects which set of parameters to use for each layer of the model's trining when traversing
