+++ worker/src/lib.rs
@@ -1,6 +1,3 @@
 pub mod builder;
-pub mod error;
 pub mod middleware;
 pub mod worker;
-
-pub use error::{Result, WorkerErr};
